+++ packages/user_preferences/local_storage_user_preferences_client/test/local_storage_user_preferences_client_test.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:local_storage_user_preferences_client/local_storage_user_preferences_client.dart';
+
+void main() {}